Add prefix search to doc search.
If not already a prefix search, we also search for the query
with a wildcard at the end (matching things that begin with it)
and append those results to the end of the list (so direct
matches have priority).

Also modify the highlight plugin to correctly handle prefix matches
(i.e. match word boundaries only if it's not a prefix match)

Resolves: LE-5375
Change-Id: I0f040d156c734ec1380803383ac88c3f39fb2406